modules/detection/requirements.txt
redis
numpy
torch==1.6.0
torchvision==0.7.0
PyTurboJPEG
